+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support associative array `set` and `get`.
 - Support `MailBox` class.
 - Support break statements.
+- Unroll loops indexing into generate for blocks.
 
 ## [0.1.14]
 ### Importer